Refactor form template handling and update placeholders in templates
